added jwt-extended

+++ resources/item.py
@@ -1,8 +1,6 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
-
-#retirvee items from a db
 
 
 class Item(Resource):
@@ -18,7 +16,7 @@ class Item(Resource):
         help="Every item needs a store id"
         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         try:
             item = ItemModel.find_by_name(name)
@@ -36,7 +34,7 @@ class Item(Resource):
             return {"message": "and item with name '{}' already exists".format(name)}, 400
 
         data = Item.parser.parse_args()
-        item = ItemModel(name, data['price'], data['store_id'])
+        item = ItemModel(name, **data)
 
         try:
             item.save_to_db()
@@ -45,8 +43,12 @@ class Item(Resource):
 
         return item.json(), 201
 
-
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required"}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -59,7 +61,7 @@ class Item(Resource):
         item = ItemModel.find_by_name(name)
 
         if item is None:
-            item = ItemModel(name, data['price'], data['store_id'])
+            item = ItemModel(name, **data)
         else:
             item.price = data['price']
 
@@ -69,4 +71,4 @@ class Item(Resource):
 
 class ItemList(Resource):
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))} #.all return all the objects in the db
+        return {'items': [x.json() for x in ItemModel.find_all()]} #.all return all the objects in the db
